requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
supabase==2.19.0
python-dotenv==1.1.1
psycopg2-binary==2.9.9
